backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Initial FastAPI backend setup for supply-chain coordinator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.18.4",
    "fastapi[standard]==0.135.1",
    "google-genai==1.68.0",
    "httpx==0.28.1",
    "langgraph==1.1.3",
    "langsmith==0.7.22",
    "opentelemetry-sdk==1.40.0",
    "psycopg==3.3.3",
    "pydantic==2.12.5",
    "redis==7.3.0",
    "resend==2.25.0",
    "slack-sdk==3.41.0",
    "sqlalchemy==2.0.48",
    "twilio==9.10.3",
    "uvicorn==0.42.0",
    "websockets==16.0",
]
